Distinct in naming service/workload
